Validacion al eliminar la imagen.

+++ app/Http/Controllers/NursesController.php
@@ -217,6 +217,12 @@ class NursesController extends Controller
         $nurse->image_path = null;
         $nurse->save();
 
+        if (!$nurse->image_path) {
+            return response()->json([
+                'message' => 'No Image Found'
+            ], 404);
+        }
+
         Storage::disk('s3')->delete($nurse->image_path);
 
         return response()->json([
